Move the file to the correct place
