+++ newclient/src/modules/tokens/handlers/tokensMarketInfo/TokensMarketInfo.vue
@@ -1,16 +1,34 @@
 <template>
     <v-card color="white" flat class="pb-2">
-        <app-table-title :title="getTitle" :has-pagination="false" :page-type="pageType" page-link="" />
+        <app-table-title
+            :title-caption="$vuetify.breakpoint.smAndDown ? $t('token.top-200') : ''"
+            :title="getTitle"
+            :has-pagination="false"
+            :page-type="pageType"
+            page-link=""
+        />
         <v-divider />
-        <v-layout v-if="showPagination && !initialLoad" row wrap align-center justify-space-between pl-3 pr-3>
-            <v-flex xs12 md4>
-                <p class="info--text">Top 200 tokens</p>
+        <v-layout v-if="!initialLoad" row wrap align-center justify-space-between pl-3 pr-3>
+            <v-flex xs12 md4 hidden-sm-and-down>
+                <p class="info--text">{{ $t('token.top-200') }}</p>
             </v-flex>
             <v-flex hidden-sm-and-down md4>
                 <!-- Search Bar -->
             </v-flex>
-            <v-flex xs12 md4>
-                <app-paginate :total="totalPages" :current-page="index" :has-input="true" :has-first="true" :has-last="true" @newPage="setPage" />
+            <v-flex xs12 sm12 md4>
+                <v-layout :align-end="$vuetify.breakpoint.mdAndUp" :align-center="$vuetify.breakpoint.smAndDown" d-flex column>
+                    <app-filter :options="options" :show-desktop="false" :is-sort="true" @onSelectChange="sortTokens" />
+                    <app-paginate
+                        v-if="showPagination"
+                        :total="totalPages"
+                        :current-page="index"
+                        :has-input="true"
+                        :has-first="true"
+                        :has-last="true"
+                        class="pb-2"
+                        @newPage="setPage"
+                    />
+                </v-layout>
             </v-flex>
         </v-layout>
         <table-txs :max-items="maxItems" :index="index" :is-loading="initialLoad" :table-message="message" :txs-data="showTokens" :is-scroll-view="false">
@@ -23,13 +41,20 @@
                 </v-card>
             </template>
         </table-txs>
-        <v-layout v-if="showPagination && !initialLoad" justify-end row class="pb-3 pr-4">
+        <v-layout
+            v-if="showPagination && !initialLoad"
+            :justify-end="$vuetify.breakpoint.mdAndUp"
+            :justify-center="$vuetify.breakpoint.smAndDown"
+            row
+            class="pb-3 pr-4"
+        >
             <app-paginate :total="totalPages" :current-page="index" :has-input="true" :has-first="true" :has-last="true" @newPage="setPage" />
         </v-layout>
     </v-card>
 </template>
 
 <script lang="ts">
+import AppFilter from '@app/core/components/ui/AppFilter.vue'
 import AppTableTitle from '@app/core/components/ui/AppTableTitle.vue'
 import AppPaginate from '@app/core/components/ui/AppPaginate.vue'
 import TableTxs from '@app/modules/txs/components/TableTxs.vue'
@@ -39,24 +64,19 @@ import { getLatestPrices_getLatestPrices as TokenMarketData } from '@app/core/co
 import { CoinData } from '@app/core/components/mixins/CoinData/CoinData.mixin'
 import { Component, Prop, Watch, Mixins } from 'vue-property-decorator'
 import BN from 'bignumber.js'
-
 const MAX_TOKENS = 200
-
 const KEY_VOLUME = 'total_volume'
 const KEY_SYMBOL = 'symbol'
 const KEY_PRICE = 'current_price'
 const KEY_MARKET_CAP = 'market_cap'
-
 interface TokensSortedInterface {
     ascend: TokenMarketData[] | null
     desend: TokenMarketData[] | null
 }
-
 class TokensSorted implements TokensSortedInterface {
     /* Properties: */
     ascend: TokenMarketData[]
     desend: TokenMarketData[]
-
     /* Constructor: */
     constructor(data: TokenMarketData[], sortKey: string) {
         this.desend = this.sortByKeyDesend([...data], sortKey)
@@ -76,21 +96,20 @@ class TokensSorted implements TokensSortedInterface {
         return this.desend
     }
 }
-
 /*
   DEV NOTES:
   - add on Error
   - add messages if Error to be displayed in Table
 */
 const FILTER_VALUES = ['name_high', 'name_low', 'price_high', 'price_low', 'volume_high', 'volume_low', 'market_cap_high', 'market_cap_low']
-
 @Component({
     components: {
         AppTableTitle,
         AppPaginate,
         TableTxs,
         TableTokensHeader,
-        TableTokensRow
+        TableTokensRow,
+        AppFilter
     }
 })
 export default class AddressTokens extends Mixins(CoinData) {
@@ -99,36 +118,75 @@ export default class AddressTokens extends Mixins(CoinData) {
       Props
     ===================================================================================
     */
-
     @Prop(Number) maxItems!: number
     @Prop(String) address!: string
     @Prop({ type: String, default: 'eth' }) tokenType!: string
-
     /*
     ===================================================================================
      Initial Data
     ===================================================================================
     */
-
     error = ''
     pageType = 'tokens'
     initialLoad = true
     index = 0
     totalPages = 3
     totalTokens = 0
-    isSortedBy = FILTER_VALUES[0]
+    isSortedBy = FILTER_VALUES[2]
     tokensData: TokenMarketData[] | null = null
     tokensByMarket!: TokensSorted
     tokensBySymbol!: TokensSorted
     tokensByPrice!: TokensSorted
     tokensByVolume!: TokensSorted
-
     /*
     ===================================================================================
       Computed
     ===================================================================================
     */
-
+    get options() {
+        return [
+            {
+                value: FILTER_VALUES[0],
+                text: this.$i18n.tc('token.name', 1),
+                filter: this.$i18n.t('filter.high')
+            },
+            {
+                value: FILTER_VALUES[1],
+                text: this.$i18n.tc('token.name', 1),
+                filter: this.$i18n.t('filter.low')
+            },
+            {
+                value: FILTER_VALUES[2],
+                text: this.$i18n.tc('price.name', 1),
+                filter: this.$i18n.t('filter.high')
+            },
+            {
+                value: FILTER_VALUES[3],
+                text: this.$i18n.tc('price.name', 1),
+                filter: this.$i18n.t('filter.low')
+            },
+            {
+                value: FILTER_VALUES[4],
+                text: this.$i18n.tc('token.volume', 1),
+                filter: this.$i18n.t('filter.high')
+            },
+            {
+                value: FILTER_VALUES[5],
+                text: this.$i18n.tc('token.volume', 1),
+                filter: this.$i18n.t('filter.low')
+            },
+            {
+                value: FILTER_VALUES[6],
+                text: this.$i18n.t('token.market'),
+                filter: this.$i18n.t('filter.high')
+            },
+            {
+                value: FILTER_VALUES[7],
+                text: this.$i18n.t('token.market'),
+                filter: this.$i18n.t('filter.low')
+            }
+        ]
+    }
     get showTokens(): TokenMarketData[] {
         const start = this.index * this.maxItems
         if (!this.initialLoad && this.tokensData) {
@@ -140,18 +198,15 @@ export default class AddressTokens extends Mixins(CoinData) {
     get showPagination(): boolean {
         return this.totalPages > 1
     }
-
     get message(): string {
         if (this.error != '') {
             return this.error
         }
         return ''
     }
-
     get getTitle(): string {
         return `${this.$tc('token.name', 2)}`
     }
-
     /*
     ===================================================================================
       Methods:
@@ -176,7 +231,6 @@ export default class AddressTokens extends Mixins(CoinData) {
             }
         }
     }
-
     setPage(page: number, reset: boolean = false): void {
         if (reset) {
             this.index = 0
@@ -186,7 +240,6 @@ export default class AddressTokens extends Mixins(CoinData) {
             this.index = page
         }
     }
-
     /*
     ===================================================================================
       Watch:
@@ -196,14 +249,14 @@ export default class AddressTokens extends Mixins(CoinData) {
     onisLoadingTokensMarketData(newVal: boolean, oldVal: boolean) {
         if (!newVal) {
             const marketData = this.getEthereumTokens()
-            // marketData sort by Price High to low , slice it to 200
             if (marketData !== false) {
-                this.tokensByVolume = new TokensSorted(marketData, KEY_VOLUME)
-                this.tokensByMarket = new TokensSorted(marketData, KEY_MARKET_CAP)
-                this.tokensBySymbol = new TokensSorted(marketData, KEY_SYMBOL)
-                this.tokensByPrice = new TokensSorted(marketData, KEY_PRICE)
+                const marketDataByPrice = new TokensSorted(marketData, KEY_PRICE).getAscend()
+                this.tokensByVolume = new TokensSorted(marketDataByPrice, KEY_VOLUME)
+                this.tokensByMarket = new TokensSorted(marketDataByPrice, KEY_MARKET_CAP)
+                this.tokensBySymbol = new TokensSorted(marketDataByPrice, KEY_SYMBOL)
+                this.tokensByPrice = new TokensSorted(marketDataByPrice, KEY_PRICE)
                 this.sortTokens(this.isSortedBy)
-                this.totalTokens = marketData.length
+                this.totalTokens = marketDataByPrice.length
                 this.totalPages = Math.ceil(new BN(this.totalTokens).div(this.maxItems).toNumber())
                 this.initialLoad = false
             } else {
@@ -211,7 +264,6 @@ export default class AddressTokens extends Mixins(CoinData) {
             }
         }
     }
-
     /*
     ===================================================================================
       LifeCycle:
